+++ extensions/raycast-notification/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Raycast Notification Changelog
 
-## Added icons and screenshots - 2024-08-08
+## [Improvements] - {PR_MERGE_DATE}
+
+- Moved screenshots to metadata folder and used raycast colors for the command icon.
+- Supported success and failure notification types with Toast API.
 
 ## [Initial Version] - 2024-08-04